backend/requirements.txt
# Core dependencies
Flask==3.0.0
gunicorn==21.2.0

# ML Libraries (updated for Python 3.12 compatibility)
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.2

# Data preprocessing
scipy>=1.11.2

# Optional: Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0

# CORS support
flask-cors==4.0.0

# Environment variables
python-dotenv==1.0.0
